+++ app/javascript/engine/actions/UndeployAction.ts
@@ -1,10 +1,10 @@
 import { Coordinate } from "../../utilities/commonTypes";
 import { formatCoordinate, formatNation } from "../../utilities/graphics";
-import DeployState from "../control/state/DeployState";
 import Game from "../Game";
 import { GameActionPath, GameActionData, GameActionReinforcementUnit } from "../GameAction";
 import Unit from "../Unit";
 import BaseAction from "./BaseAction";
+import StackingActionError from "./StackingActionError";
 
 export default class UndeployAction extends BaseAction {
   target: Coordinate;
@@ -12,6 +12,7 @@ export default class UndeployAction extends BaseAction {
   rKey: string;
   rId: string;
   rName: string;
+  index: number;
 
   constructor(data: GameActionData, game: Game, index: number) {
     super(data, game, index)
@@ -28,6 +29,7 @@ export default class UndeployAction extends BaseAction {
     this.rTurn = deploy[0].turn
     this.rId = deploy[0].id
     this.rName = deploy[0].name
+    this.index = deploy[0].index ?? 0
   }
 
   get type(): string { return "undeploy" }
@@ -48,7 +50,28 @@ export default class UndeployAction extends BaseAction {
       scenario.replaceAxisReinforcement(this.rTurn, this.rKey)
     }
     this.map.removeCounter(this.target, this.rId)
-    this.game.setGameState(new DeployState(this.game, this.rTurn, this.rKey))
+    const units = this.map.units[this.target.y][this.target.x]
+    if (units.length > this.index) {
+      const u = units[this.index] as Unit
+      const last = this.index > 0 ? units[this.index - 1] as Unit : undefined
+      if (u.uncrewedSW) {
+        if (!last || !last.canCarrySupport) {
+          throw new StackingActionError(
+            `${u.name} is no longer assigned to an operator; it ` +
+              "must be stacked on a squad, team, or leader to be assigned."
+          )
+        }
+      } else if (u.crewed) {
+        if (!last || !last.canHandle && !(last.canTow && last.size >= (u.towSize ?? 0))) {
+          throw new StackingActionError(
+            `${u.name} is no longer assigned to an operator or vehicle; it ` +
+              "must be placed on a squad or team to be assigned, or on a vehicle large enough " +
+              "to tow it."
+          )
+
+        }
+      }
+    }
   }
 
   undo(): void {
@@ -63,6 +86,6 @@ export default class UndeployAction extends BaseAction {
       (uf as Unit).playerNation = this.player === 1 ? scenario.alliedFactions[0] : scenario.axisFactions[0]
     }
     uf.id = this.rId
-    map.addCounter(this.target, uf)
+    map.addCounter(this.target, uf, this.index)
   }
 }